+++ app/src/main/java/com/nagyrobi144/wearable/hrv/feature/tiles/HrvTileState.kt
@@ -1,3 +1,3 @@
 package com.nagyrobi144.wearable.hrv.feature.tiles
 
-data class HrvTileState(val averageHrv: Double, val minHrv: Int, val maxHrv: Int)
+data class HrvTileState(val averageHrv: Int, val minHrv: Int, val maxHrv: Int)
